scraper/pyproject.toml
[tool.poetry]
name = "scraper"
version = "0.1.0"
description = ""
authors = ["StigHaraldGustavsen <stighg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
selenium = "4.9.1"
beautifulsoup4 = "4.12.2"
webdriver-manager = "4.0.0"
lxml = "^5.1.0"
python-dotenv = "^1.0.1"
flask = "^3.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
